LayoutTests/fast/selectors/specificity-overflow-2.html

<!DOCTYPE html>
<head>
    <style>
        /* Saturate A and B fields: (255, 255, 0). */
        .target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target {
            color: rgb(1, 2, 3);
        }

        /* Saturate B and C fields: (0, 255, 255). */
        target:not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar).target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target {
            background-color: rgb(4, 5, 6);
        }

        /* Saturate A and C fields: (255, 0, 255). */
        target:not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar)#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target {
            margin: 25px;
        }

        /* Saturate everything: (255, 255, 255). */
        target:not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar).target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target {
            padding: 25px;
        }

        /* Overflow the A and B fields, the right output should be saturating to (255, 255, 0). Due to the position, this should override the previous (255, 255, 0). */
        .target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target:not(.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar.foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar#foobar)#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target {
            color: rgb(7, 8, 9);
        }

        /* Overflow the B and C fields, the right output should be saturating to (0, 255, 255). */
        target:not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar).target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target:not(:not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target).target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target) {
            background-color: rgb(10, 11, 12);
        }

        /* Overflow A and C fields. */
        target:not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar)#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target:not(:not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target)#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target) {
            margin: 0;
        }

        /* -Overflow everyone.
           -What do you mean "everyone"?
           -EVERYONE!
         */
        target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target:not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar):not(foobar)#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target:not(:not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target):not(target).target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target.target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target#target) {
            padding: 0;
        }

    </style>
    <script src='../../resources/js-test-pre.js'></script>
</head>
<body>
    <target id="target" class="target"></target>
    <script>
    description("Verify that saturating multiple component of the specificity works correctly.");

    shouldBeEqualToString("getComputedStyle(document.getElementById('target')).color", "rgb(7, 8, 9)");
    shouldBeEqualToString("getComputedStyle(document.getElementById('target')).backgroundColor", "rgb(10, 11, 12)");
    shouldBeEqualToString("getComputedStyle(document.getElementById('target')).margin", "0px");
    shouldBeEqualToString("getComputedStyle(document.getElementById('target')).padding", "0px");
    </script>
<script src='../../resources/js-test-post.js'></script>
</body>
