Safari-611.1.10/LayoutTests/fast/selectors/specificity-overflow-2-expected.txt

Verify that saturating multiple component of the specificity works correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.getElementById('target')).color is "rgb(7, 8, 9)"
PASS getComputedStyle(document.getElementById('target')).backgroundColor is "rgb(10, 11, 12)"
PASS getComputedStyle(document.getElementById('target')).margin is "0px"
PASS getComputedStyle(document.getElementById('target')).padding is "0px"
PASS successfullyParsed is true

TEST COMPLETE
